--feat: Extract inventory settings sections into feature modules
